# swyftr-swagger.yaml
swagger: '2.0'
info:
  version: '1'
  title: Swyftr
  description: Backend API for Swyftr system
host: 52.221.58.5
basePath: /api/v1
schemes:
  - https
...
paths:
  /auth/cus/register:
    post:
      summary: Register a customer to Swyftr
      tags:
        - Customer
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RegisterUserRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/RegisterUserResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/RegisterUserResponse'
  /auth/sys/register:
    post:
      summary: Add a system to Swyftr
      tags:
        - System
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              addedByUserId:
                type: string
              addedToOrgId:
                type: string
              addedToBrach:
                type: string
              employeeId:
                type: string
              userPermissions:
                $ref: '#/definitions/PermissionSet'
              userDetails:
                $ref: '#/definitions/RegisterUserRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/RegisterUserResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/RegisterUserResponse'
...
  /auth/cus/login:
    post:
      summary: Login a customer
      tags:
        - Customer
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CusLoginResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: OTP Not Verified
        '500':
          description: Fail
  /auth/sys/login:
    post:
      summary: Login a system user
      tags:
        - System
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SysLoginResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Fail
  /auth/rider/login:
    post:
      summary: Login a rider
      tags:
        - Rider
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/RiderLoginResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Fail
  /auth/cus/send_reset_password:
    post:
      summary: Send reset password link to user's email address
      tags:
        - Sprint1
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Fail
  /auth/cus/reset_password:
    post:
      summary: Update customer password using the token recieved to his email
      tags:
        - Customer
        - Auth
        - Sprint1
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              token:
                type: string
              newPassword:
                type: string
              retypePassword:
                type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Fail
...
  /auth/sys/permission:
    post:
      summary: Grant Permissions For A User
      tags:
        - Sprint2
        - Auth
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: integer
              permissions:
                type: array
                items:
                  $ref: '#/definitions/Permission'
      responses:
        200:
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Fail
...
  /category/cus/allSubCategories/{categoryId}:
    get:
      summary: Get all categories belongs to a parent category
      tags:
        - Sprint2
        - Category
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: categoryId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              subSubCategories:
                type: array
                items:
                  $ref: '#/definitions/Category'
  /category/allSubCategories/{categoryId}:
    get:
      summary: Get all categories belongs to a parent category
      tags:
        - Sprint2
        - Category
        - System
      produces:
        - application/json
      parameters:
        - in: path
          name: categoryId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              subSubCategories:
                type: array
                items:
                  $ref: '#/definitions/Category'
  /category/cus/allStores/{categoryId}:
    get:
      summary: Get all stores belongs to a category
      tags:
        - Sprint2
        - Category
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: categoryId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              stores:
                type: array
                items:
                  $ref: '#/definitions/StoreWithOrg'
              partners:
                type: array
                items:
                  $ref: '#/definitions/Organization'
  /category/allStores/{categoryId}:
    get:
      summary: Get all stores belongs to a category
      tags:
        - Sprint2
        - Category
        - System
      produces:
        - application/json
      parameters:
        - in: path
          name: categoryId
          type: number
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              stores:
                type: array
                items:
                  $ref: '#/definitions/StoreWithOrg'
              partners:
                type: array
                items:
                  $ref: '#/definitions/Organization'
...
definitions:
  LoginRequest:
    type: object
    properties:
      userName:
        type: string
      password:
        type: string
  CusLoginResponse:
    type: object
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      expiresOn:
        type: integer
      userId:
        type: string
  RiderLoginResponse:
    type: object
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      expiresOn:
        type: integer
      userId:
        type: string
  SysLoginResponse:
    type: object
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      expiresOn:
        type: integer
      userId:
        type: string
      permissions:
        $ref: '#/definitions/PermissionSet'
  RegisterUserRequest:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      phoneNo:
        type: string
      mobileNo:
        type: string
      nic:
        type: string
      password:
        type: string
      retypePassword:
        type: string
      profilePicId:
        type: number
  RegisterUserResponse:
    type: object
    properties:
      status:
        type: number
      message:
        type: number
  PermissionSet:
    type: array
    items:
      type: string
...
  Organization:
    type: object
    properties:
      orgId:
        type: string
      name:
        type: string
      headOfficeAddress:
        type: string
      headOfficeContactNo:
        type: string
      orgType:
        type: string
        enum: ['swyftr', 'partner_store', 'partner_rider']
      bank:
        type: string
      branch:
        type: string
      accountNo:
        type: string
      preferedPaymentMethod:
        type: string
        enum: ['cash', 'cheque']
      branchCode:
        type: string
      bankSwiftCode:
        type: string
      pointOfContactName:
        type: string
      pointOfContactMobile:
        type: string
      tags:
        type: string
      orgLogoImage:
        type: string
...
  Permission:
    type: object
    properties:
      permissionId:
        type: integer
      allowed:
        type: number
...
  StoreWithOrg:
    type: object
    properties:
      id:
        type: integer
      orgId:
        type: integer
      orgLogoId:
        type: integer
      branchName:
        type: string
      address:
        type: string
      longitude:
        type: number
      latitude:
        type: number
      district:
        type: string
      province:
        type: string
      managerId:
        type: integer
      phone:
        type: string
      logo:
        type: number
      displayImage:
        type: number
      slogan:
        type: string
      branchLogoImage:
        type: string
...
  Category:
    type: object
    properties:
      id:
        type: integer
      categoryName:
        type: string
      parentCategory:
        type: string
      assetId:
        type: number
...